refactor(page): ♻️  use modal via ui context
